@@ -3,7 +3,9 @@
   "version": "1.0.0",
   "description": "REST API para o backend do Tuitui",
   "main": "src/server.ts",
-  "engines": { "node": "12.18.3" },
+  "engines": {
+    "node": ">=16"
+  },
   "scripts": {
     "start": "node dist/src/server.js",
     "dev": "tsnd --transpile-only --ignore-watch node_modules --respawn src/server.ts",
@@ -45,6 +47,9 @@
     "@types/jest": "^26.0.15",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/mongoose": "^5.10.0",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "@typescript-eslint/parser": "^5.41.0",
+    "eslint": "^8.26.0",
     "factory-girl": "^5.0.4",
     "faker": "^5.1.0",
     "jest": "^26.6.3",
