@@ -1,20 +1,27 @@
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
+import { NextFunction } from "express";
 
-import { NextFunction } from 'express';
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
 
-const { Schema, model } = require('../../database');
+const { Schema, model } = require("../../database");
 
 const UserSchema = new Schema({
   username: {
     type: String,
     required: true,
-    unique: true,
+    unique: true
   },
   password: {
     type: String,
     required: true,
-    select: false,
+    select: false
+  },
+  description: {
+    type: String
+  },
+  userEmail: {
+    type: String,
+    required: true
   },
   createdAt: {
     type: Date,
@@ -22,24 +29,24 @@ const UserSchema = new Schema({
   },
 });
 
-UserSchema.pre('save', async function(this: typeof UserSchema, next: NextFunction) {
+UserSchema.pre("save", async function(this: typeof UserSchema, next: NextFunction) {
   const hash = await bcrypt.hash(this.password, 10);
   this.password = hash;
   
   next();
 });
 
-const User = model('User', UserSchema);
+const User = model("User", UserSchema);
 
 User.prototype.generateToken = function () {
   return jwt.sign({ id: this._id }, process.env.SECRET_KEY, {
     expiresIn: 86400,
   });
-}
+};
 
 User.prototype.checkPassword = function(password: string) {
   return bcrypt.compare(password, this.password);
-}
+};
 
 module.exports = User;
 
